Use query-circuitBreaker BigArrays for query execution
The injected BigArrays instance uses the Request circuit breaker (for
accounting only; it doesn't trip the breaker)

For query execution we want to use the query circuit breaker and we also
want it to trip the breaker. E.g. in the `GroupByOptimizedIterator` we
create a `BigObjectArray` instance to hold all the keys. This should
trip eagerly if under memory pressure.